requirements.txt
fastapi
uvicorn
pillow
python-multipart
requests